Fixed scaling problem with solar input power. Oops.
I found a problem in my scaling of the PV Input Power data. The documentation from Victron is a bit confusing; the units say "1W" but then the Remarks say "un32 @ 0.01W". The first appears to be correct, so no scaling is needed.

I also updated the comments at the top of the code with a pointer to Victron's documentation on the beacons - and a thank you!